third_party/WebKit/LayoutTests/fast/events/autoscroll-disabled-user-select-none.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="container" style="height: 400px; overflow: auto;">
  <div style="-webkit-user-select:none; height: 2000px; background-color: yellow">
    Hello World! Hello Woorld! Hello Wooorld! Hello Woooorld!<br />
    Hello World! Hello Woorld! Hello Wooorld! Hello Woooorld!<br />
    Hello World! Hello Woorld! Hello Wooorld! Hello Woooorld!<br />
    Hello World! Hello Woorld! Hello Wooorld! Hello Woooorld!<br />
    Hello World! Hello Woorld! Hello Wooorld! Hello Woooorld!<br />
  </div>
</div>
<script>
var testSelectNone = async_test("Selection-autoscroll should not be triggered when the selection happens in an element with user-select:none");
testSelectNone.step(function() {
    if (!window.eventSender)
        return;
    var dragStartX = 50;
    var dragStartY = 50;
    var dragEndX = 60;
    var dragEndY = 400;
    var container = document.getElementById("container");

    eventSender.dragMode = false;
    eventSender.mouseMoveTo(dragStartX, dragStartY);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(dragEndX, dragEndY);

    requestAnimationFrame(function() {
        assert_equals(container.scrollTop, 0);
        testSelectNone.done();
    });
});
</script>
